Add 'private' flag to incidents
Allows marking incidents to signal they contain private information.
This can be done on creation or later when editing the incident.
The incident summary, impact, severity etc. can still be edited after
an incident is made private.

Marking an incident as private does not automatically change the comms
channel privacy settings, this needs to be done manually.

Once the flag is set to true it cannot be undone from Slack.